pyproject.toml
[project]
name = "hanzi-animator"
version = "0.1.0"
description = "蒙版法汉字书写动画生成系统"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "imageio>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "ruff>=0.0.287",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
